tg-cipher-bot/requirements.txt
aiogram==3.5.0
pytest==8.2.0
pytest-asyncio==0.23.6
coverage==7.5.1
coverage-badge==1.1.0
python-dotenv==1.0.1